Added user constrains and changed type of baseSalary from float to double

+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -24,7 +24,7 @@ public class User {
     private String password;
     private String phone;
     private String address;
-    private float baseSalary;
+    private double baseSalary;
     private String roleId;
 
 }
